General/global: A first set of changes towards better LLP64 compatibility (besides existing ILP32 and LP64 compatibility).
git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@98 5fe0b3f3-48df-4727-b460-e6dda9226989

+++ Libs/String.hpp
@@ -33,8 +33,8 @@ namespace cf
     {
         inline bool EndsWith(const std::string& String, const std::string& Suffix)
         {
-            const int StringLength=String.length();
-            const int SuffixLength=Suffix.length();
+            const size_t StringLength=String.length();
+            const size_t SuffixLength=Suffix.length();
 
             if (StringLength<SuffixLength) return false;
